requirements.txt
boto3==1.34.2
google-cloud-storage==2.14.0
PyYAML==6.0.1
pytest==7.4.3
